Frausar_API_GUI/requirements.txt
fastapi
uvicorn
pandas
numpy
openpyxl
pyyaml
requests
ruamel.yaml
pydantic 